requirements.txt
aiocontextvars==0.2.2
apturl==0.5.2
asgiref==3.3.1
asn1crypto==0.24.0
bottle==0.12.13
Brlapi==0.6.6
catfish==1.4.4
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
contextvars==2.4
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
dataclasses==0.8
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-utils-six==2.0
djangorestframework==3.12.4
docker==2.5.1
docker-pycreds==0.2.1
Glances==2.11.1
httplib2==0.9.2
idna==2.6
immutables==0.15
influxdb==4.1.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
loguru==0.5.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mate-hud==18.4.8
mate-tweak==18.4.16
matplotlib==2.1.1
menulibre==2.2.0
mugshot==0.4.0
name-that-hash==1.4.0
netifaces==0.10.4
numpy==1.13.3
oauth==1.0.1
olefile==0.45.1
onboard==1.4.1
pexpect==4.2.1
Pillow==5.1.0
ply==3.11
protobuf==3.0.0
psutil==5.4.2
psycopg2==2.8.6
pyasn1==0.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycryptodomex==3.4.7
pycups==1.9.73
pycurl==7.43.0.1
Pygments==2.8.1
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
pysmi==0.2.2
pysnmp==4.4.3
pystache==0.5.4
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.6.1
python-debian==0.1.32
python-debianbts==2.7.2
python-magic==0.4.16
python-xapp==1.0.0
python-xlib==0.20
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
rich==9.13.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
setproctitle==1.1.10
sgt-launcher==0.2.4
simplejson==3.13.2
six==1.11.0
sqlparse==0.4.1
ssh-import-id==5.7
system-service==0.3
systemd-python==234
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
websocket-client==0.44.0
xcffib==0.5.1
xkit==0.0.0
zope.interface==4.3.2
